train on original set

+++ pytorchtest.py
@@ -83,7 +83,7 @@ def train_model(model, criterion, optimizer, lr_scheduler):
                     optimizer.step()
                 # statistics
                 running_loss += loss.data[0]
-                running_corrects += torch.sum(preds == labels.data.int())
+                running_corrects += torch.sum(preds == labels.data.long())
                 # next step
                 step += 1
             epoch_loss = running_loss / dset_sizes[phase]
